requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-OAuthlib==0.9.6
Flask-CORS==4.0.0
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==41.0.7
python-multipart==0.0.6
validators==0.22.0 